matchingapp-backend/matchingapp/requirements.txt
django
djangorestframework
django-cors-headers
python-dotenv
dj-database-url
psycopg2-binary
djangorestframework-simplejwt
djangorestframework-jwt
django-allauth
dj-rest-auth
django-rest-auth
django-debug-toolbar
django-extensions
djangorestframework-jsonapi
pyjwt
drf-yasg
Pillow
python-decouple
numpy
pandas
scikit-learn
